+++ src/remedy/interfaces/routes/__init__.py
@@ -8,6 +8,7 @@ from remedy.interfaces.routes.auth import register_auth_routes
 from remedy.interfaces.routes.catalog import register_catalog_routes
 from remedy.interfaces.routes.chat import register_chat_routes
 from remedy.interfaces.routes.computer import register_computer_routes
+from remedy.interfaces.routes.hive import register_hive_routes
 from remedy.interfaces.routes.memory import register_memory_routes
 from remedy.interfaces.routes.misc import register_misc_routes
 from remedy.interfaces.routes.nanoswarm import register_nanoswarm_routes
@@ -53,5 +54,6 @@ def register_all_routes(
     register_telephony_routes(app, **kw)
     register_rmb_routes(app, **kw)
     register_nanoswarm_routes(app, **kw)
+    register_hive_routes(app, **kw)
     register_usage_routes(app, **kw)
     register_webhook_routes(app, **kw)